+++ src/app/organisations/organisation/organisation.component.ts
@@ -1,11 +1,12 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, Input, Output, EventEmitter, OnInit} from '@angular/core';
 import {OrganisationEntityService} from '../services/organisation-entity.service';
 import {ActivatedRoute, Router} from '@angular/router';
-import {map, withLatestFrom} from 'rxjs/operators';
-import {Observable} from 'rxjs';
+import {map} from 'rxjs/operators';
+import {Observable, combineLatest} from 'rxjs';
 import {Organisation} from '../model/organisation';
-import {MessageService} from 'primeng/api';
+import {ConfirmationService, MessageService} from 'primeng/api';
 import {enmStatusCompany, enmGender, enmCountry} from '../../shared/enums';
+import {NgForm} from '@angular/forms';
 
 @Component({
   selector: 'app-organisation',
@@ -15,7 +16,11 @@ import {enmStatusCompany, enmGender, enmCountry} from '../../shared/enums';
 export class OrganisationComponent implements OnInit {
 
   @Input() idDis$: Observable<number>;
-  organisation$: Observable<Organisation>;
+    @Output() onOrganisationUpdate = new EventEmitter<Organisation>();
+    @Output() onOrganisationDelete = new EventEmitter<Organisation>();
+    @Output() onOrganisationQuit = new EventEmitter<Organisation>();
+    booCanDeleteAndQuit: boolean;
+  organisation: Organisation;
   genders: any[];
   statuts: any[];
   countries: any[];
@@ -24,53 +29,81 @@ export class OrganisationComponent implements OnInit {
       private organisationsService: OrganisationEntityService,
       private route: ActivatedRoute,
       private router: Router,
-      private messageService: MessageService
+      private messageService: MessageService,
+      private confirmationService: ConfirmationService
   ) {
       this.statuts = enmStatusCompany;
       this.genders = enmGender;
       this.countries = enmCountry;
+      this.booCanDeleteAndQuit = true;
   }
 
   ngOnInit(): void {
 // comment: this component is sometimes called from his parent Component with BankId @Input Decorator,
       // or sometimes via a router link via the Main Menu
-      console.log('Statuts:', this.statuts);
-      if (!this.idDis$) {
+         if (!this.idDis$) {
           // we must come from the menu
           console.log('We initialize a new organisation object from the router!');
+          this.booCanDeleteAndQuit = false;
           this.idDis$ = this.route.paramMap
               .pipe(
                   map(paramMap => paramMap.get('idDis')),
                   map(idDisString => Number(idDisString))
               );
       }
-      this.organisation$ = this.idDis$
+      const organisation$ = combineLatest([this.idDis$, this.organisationsService.entities$])
           .pipe(
-              withLatestFrom(this.organisationsService.entities$),
-              map(([idDis, organisations]) => {
-                  console.log('finding organisation with id', idDis);
-                  return organisations.find(organisation => idDis === organisation.idDis  );
-
-              })
+              map(([idDis, organisations]) => organisations.find(organisation => idDis === organisation.idDis))
           );
 
+      organisation$.subscribe(organisation => {
+          this.organisation = organisation;
+          console.log('Organisation : ', organisation);
+      });
   }
-  delete(organisation: Organisation) {
-    const  myMessage = {severity: 'success', summary: 'Destruction', detail: `L'organisation ${organisation.idDis} ${organisation.societe}  a été détruite`};
-    this.organisationsService.delete(organisation)
-        .subscribe( ()  => {
-            this.messageService.add(myMessage);
-        });
-  }
-
   save(oldOrganisation: Organisation, organisationForm: Organisation) {
     const modifiedOrganisation = Object.assign({}, oldOrganisation, organisationForm);
     this.organisationsService.update(modifiedOrganisation)
         .subscribe( ()  => {
           this.messageService.add({severity: 'success', summary: 'Mise à jour', detail: `L'organisation ${modifiedOrganisation.idDis} ${modifiedOrganisation.societe}  a été modifiée`});
          });
-
-
   }
-
+    delete(event: Event, organisation: Organisation) {
+        this.confirmationService.confirm({
+            target: event.target,
+            message: 'Confirm Deletion?',
+            icon: 'pi pi-exclamation-triangle',
+            accept: () => {
+                const  myMessage = {severity: 'success', summary: 'Destruction', detail: `L' organisation ${organisation.societe} a été détruite`};
+                this.organisationsService.delete(organisation)
+                    .subscribe( () => {
+                        this.messageService.add(myMessage);
+                        this.onOrganisationDelete.emit();
+                    });
+            },
+            reject: () => {
+                console.log('We do nothing');
+            }
+        });
+    }
+    quit(event: Event, oldOrganisation: Organisation, organisationForm: NgForm, formDirty: boolean) {
+        if (formDirty) {
+            this.confirmationService.confirm({
+                target: event.target,
+                message: 'Your changes may be lost. Are you sure that you want to proceed?',
+                icon: 'pi pi-exclamation-triangle',
+                accept: () => {
+                    organisationForm.reset( oldOrganisation); // reset in-memory object for next open
+                    console.log('We have reset the form to its original value');
+                    this.onOrganisationQuit.emit();
+                },
+                reject: () => {
+                    console.log('We do nothing');
+                }
+            });
+        } else {
+            console.log('Form is not dirty, closing');
+            this.onOrganisationQuit.emit();
+        }
+    }
 }
